requirements.txt
freezegun>=0.3.6
pytest~=3.0.0
sqlalchemy>=1.1.0, <1.2
testing.postgresql~=1.3.0
